requirements.txt
Pillow>=10.0.0
anthropic>=0.40.0
flask>=3.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
